requirements.txt
flask
chromadb
fastembed
openai
PyPDF2
python-dotenv
requests 